ansible/requirements.txt
ansible>=2.13
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
